Operations

Operations

The following operations and transformations are provided to work on Spectra

Extinction

Spectra.extinct — Function.

extinct(::Spectrum, Av::Real, Rv::Real=3.1; law=ccm89)

Extinct a spectrum given the total extinction Av and the relative attenuation Rv. law must be a function with signature law(wave, Rv), by default we use ccm89 from DustExtinction.jl .

Examples

julia> using Spectra

julia> spec = Spectrum(range(1e4, 3e4, length=1000), randn(1000) .+ 100);

julia> extincted_spec = extinct(spec, 0.3);

julia> extinct!(spec, 0.3);

julia> spec.flux ≈ extincted_spec.flux
true

source

Spectra.extinct! — Function.

extinct!(::Spectrum, Av::Real, Rv::Real=3.1; law=ccm89)

In-place version of extinct

See Also

extinct

source