Spectrum

Spectrum

Here we will go over the Spectrum type and how we can use it to interact with spectra.

The Spectrum Type

Basic operations

Spectra.Spectrum — Type.

Spectrum(wave, flux, [sigma]; name="")

A signle dimensional astronomical spectrum. If no sigma are provided, they are assumed to be unity. The name is an optional identifier for the Spectrum. Note that the dimensions of each array must be equal or an error will be thrown.

Examples

julia> using Spectra

julia> wave = range(1e4, 4e4, length=1000);

julia> flux = randn(size(wave));

julia> spec = Spectrum(wave, flux)
Spectrum:

julia> spec = Spectrum(wave, flux, name="Just Noise")
Spectrum: Just Noise

There is easy integration with Unitful and its sub-projects

julia> using Spectra, Unitful, UnitfulAstro

julia> wave = range(1u"μm", 4u"μm", length=1000) .|> u"angstrom";

julia> sigma = randn(size(wave))u"erg/cm^2/s/angstrom";

julia> flux = sigma .+ 100u"W/m^2/m"; # There will be implicit unit promotion

julia> unit(flux[1])
kg m^-1 s^-3

julia> spec = Spectrum(wave, flux, sigma, name="Unitful")
Spectrum: Unitful

source

Base.length — Function.

length(::Spectrum)

source

Base.size — Function.

size(::Spectrum)

source

Arithmetic

The following basic arithmetic is provided

Function	Notes

+(::Spectrum, A)

-(::Spectrum, A)

*(::Spectrum, A)	sigma updated by $\sigma \left\lvert A \right\rvert$

/(::Spectrum, A)	sigma updated by $\sigma \left\lvert A \right\rvert^{-1}$

Unitful helpers

Unitful.unit — Function.

Unitful.unit(::Spectrum)

Get the units of a spectrum. Returns a tuple of the wavelength units and flux/sigma units

Examples

julia> using Spectra, Unitful, UnitfulAstro

julia> wave = range(1e4, 3e4, length=1000) |> collect;

julia> flux = wave .* 10 .+ randn(1000);

julia> spec = Spectrum(wave * u"angstrom", flux * u"W/m^2/angstrom");

julia> w_unit, f_unit = unit(spec)
(Å, W Å^-1 m^-2)

source

Unitful.ustrip — Function.

Unitful.ustrip(::Spectrum)

Remove the units from a spectrum. Useful for processing spectra in tools that don't play nicely with Unitful.jl

Examples

julia> using Spectra, Unitful, UnitfulAstro

julia> wave = range(1e4, 3e4, length=1000) |> collect;

julia> flux = wave .* 10 .+ randn(1000);

julia> spec = Spectrum(wave*u"angstrom", flux*u"W/m^2/angstrom");

julia> strip_spec = ustrip(spec);

source

Index

Spectra.Spectrum

Base.length

Base.size

Spectra.extinct

Spectra.extinct!

Spectra.resample

Spectra.resample!

Unitful.unit

Unitful.ustrip