Spectrum

Spectrum

Here we will go over the Spectrum type and how we can use it to interact with spectra.

The Spectrum Type

Spectra.Spectrum — Type.

Spectrum(wave, flux, [sigma]; name="")

A signle dimensional astronomical spectrum. If no sigma are provided, they are assumed to be unity. The name is an optional identifier for the Spectrum. Note that the dimensions of each array must be equal or an error will be thrown.

Examples

julia> using Spectra

julia> wave = range(1e4, 4e4, length=1000);

julia> flux = randn(size(wave));

julia> spec = Spectrum(wave, flux)
Spectrum:

julia> spec = Spectrum(wave, flux, name="Just Noise")
Spectrum: Just Noise

There is easy integration with $Unitful$ and its sub-projects

julia> using Spectra, Unitful, UnitfulAstro

julia> wave = range(1u"μm", 4u"μm", length=1000) .|> u"angstrom";

julia> sigma = randn(size(wave))u"erg/cm^2/s/angstrom";

julia> flux = sigma .+ 100u"W/m^2/m"; # There will be implicit unit promotion

julia> unit(flux[1])
kg m^-1 s^-3

julia> spec = Spectrum(wave, flux, sigma, name="Unitful")
Spectrum: Unitful

source

This provides the following standard library overrides

Base.length — Function.

length(::Spectrum)

source

Base.size — Function.

size(::Spectrum)

source

Index

Spectra.Spectrum

Base.length

Base.size

Spectra.extinct

Spectra.extinct!